+++ src/views/peptide/peptide_order.vue
@@ -217,7 +217,7 @@ export default {
   methods: {
     moment,
     setColumns () {
-      const { formatter } = this.$units;
+      const { formatter } = this.$utils;
       const { peptide } = this.$store.state;
       this.status = peptide.status;
       const columns = [